# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SGE Platform API
  version: 1.0.0
  description: |
    Sistema de Gestión Empresarial (SGE) — REST API for the multi-tenant
    institutional management platform. Every response follows the
    standardised envelope contract (`success: true, data, requestId` or
    `success: false, error: { code, message, details?, stage? }, requestId`).
    Schemas reflect the runtime validation; no field is documented unless
    the route handler actually returns it.
  contact:
    name: SGE Platform Team
    email: platform@cni.example.test
  license:
    name: Proprietary
servers:
  - url: https://app.cni.example.test/api
    description: Production
  - url: https://staging.cni.example.test/api
    description: Staging
  - url: http://localhost:3000/api
    description: Local development
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /audit-logs:
    get:
      tags: [Audits]
      summary: List audit log records
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: module
          in: query
          schema: { type: string }
        - name: action
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Audit log list (tenant-scoped).
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...